+++ src/realm/descriptor.hpp
@@ -293,6 +293,8 @@ public:
     /// \param link_type The type of links the column should store.
     void set_link_type(size_t col_ndx, LinkType link_type);
 
+    Spec* get_spec() const;
+
     //@{
     /// Get the descriptor for the specified subtable column.
     ///