+++ src/realm/index_string.cpp
@@ -334,6 +334,11 @@ size_t IndexArray::index_string_count(StringData value, ColumnBase* column) cons
     return index_string<index_Count, StringData>(value, dummy1, dummy2, column);
 }
 
+ref_type StringIndex::create_empty(Allocator& alloc)
+{
+    return StringIndex(nullptr, alloc).get_ref(); // Throws
+}
+
 IndexArray* StringIndex::create_node(Allocator& alloc, bool is_leaf)
 {
     Array::Type type = is_leaf ? Array::type_HasRefs : Array::type_InnerBptreeNode;
@@ -1414,7 +1419,7 @@ void StringIndex::to_dot_2(std::ostream& out, StringData title) const
     ref_type ref = get_ref();
 
     out << "subgraph cluster_string_index" << ref << " {" << std::endl;
-    out << " label = \"String index";
+    out << " label = \"Search index";
     if (title.size() != 0)
         out << "\\n'" << title << "'";
     out << "\";" << std::endl;