+++ src/realm/table.hpp
@@ -222,29 +222,34 @@ public:
 
     //@{
 
+    /// Following three search index methods take an optional `subdesc` argument.
+    /// If it is omitted, then the methods operate on `this` table. If a subtable
+    /// descriptor is passed, the methods operate on its associated subtable. NOTE: 
+    /// Subtables of the deprecated Mixed column are not supported.
+    ///
+    /// add_search_index() and remove_search_index() will add or remove search 
+    /// indexes of *all* subtables of the subtable column. This may take a while
+    /// if there are many subtables with many rows each.
+    ///
     /// has_search_index() returns true if, and only if a search index has been
     /// added to the specified column. Rather than throwing, it returns false if
     /// the table accessor is detached or the specified index is out of range.
     ///
-    /// add_search_index() adds a search index to the specified column of this
+    /// add_search_index() adds a search index to the specified column of the
     /// table. It has no effect if a search index has already been added to the
     /// specified column (idempotency).
     ///
     /// remove_search_index() removes the search index from the specified column
-    /// of this table. It has no effect if the specified column has no search
+    /// of the table. It has no effect if the specified column has no search
     /// index. The search index cannot be removed from the primary key of a
     /// table.
     ///
-    /// This table must be a root table; that is, it must have an independent
-    /// descriptor. Freestanding tables, group-level tables, and subtables in a
-    /// column of type 'mixed' are all examples of root tables. See add_column()
-    /// for more on this.
-    ///
-    /// \param column_ndx The index of a column of this table.
+    /// \param column_ndx The index of a column of the table.
 
-    bool has_search_index(size_t column_ndx) const noexcept;
-    void add_search_index(size_t column_ndx);
-    void remove_search_index(size_t column_ndx);
+   // bool has_search_index(size_t column_ndx, DescriptorRef* subdesc = nullptr) const;
+    bool has_search_index(size_t column_ndx, ConstDescriptorRef subdesc = nullptr) const;
+    void add_search_index(size_t column_ndx, DescriptorRef* subdesc = nullptr);
+    void remove_search_index(size_t column_ndx, DescriptorRef* subdesc = nullptr);
 
     //@}